# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for VibeGlobally Virtual Assistance Services
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  message: string;
}

export interface AuthMeResponse {
  authenticated: boolean;
  username?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
}

export type ContactStatus = (typeof ContactStatus)[keyof typeof ContactStatus];

export const ContactStatus = {
  new: "new",
  read: "read",
  responded: "responded",
} as const;

export interface Contact {
  id: string;
  name: string;
  email: string;
  phone?: string | null;
  company?: string | null;
  service?: string | null;
  message: string;
  status: ContactStatus;
  createdAt: string;
}

export interface CreateContactBody {
  name: string;
  email: string;
  phone?: string | null;
  company?: string | null;
  service?: string | null;
  message: string;
}

export type UpdateContactBodyStatus =
  (typeof UpdateContactBodyStatus)[keyof typeof UpdateContactBodyStatus];

export const UpdateContactBodyStatus = {
  new: "new",
  read: "read",
  responded: "responded",
} as const;

export interface UpdateContactBody {
  status?: UpdateContactBodyStatus;
}

export interface ContactList {
  contacts: Contact[];
  total: number;
  page?: number;
  limit?: number;
}

export interface Testimonial {
  id: string;
  clientName: string;
  company?: string | null;
  content: string;
  rating: number;
  facebookUrl?: string | null;
  createdAt: string;
}

export interface CreateTestimonialBody {
  clientName: string;
  company?: string | null;
  content: string;
  rating: number;
  facebookUrl?: string | null;
}

export interface TestimonialList {
  testimonials: Testimonial[];
}

export interface DashboardStats {
  totalContacts: number;
  newContacts: number;
  respondedContacts: number;
  totalTestimonials: number;
  recentContacts: Contact[];
}

export type SiteContentResponseContent = { [key: string]: unknown };

export interface SiteContentResponse {
  section: string;
  content: SiteContentResponseContent;
}

export type ListContactsParams = {
  page?: number;
  limit?: number;
  status?: ListContactsStatus;
};

export type ListContactsStatus =
  (typeof ListContactsStatus)[keyof typeof ListContactsStatus];

export const ListContactsStatus = {
  new: "new",
  read: "read",
  responded: "responded",
} as const;

export type UpdateSiteContentBody = { [key: string]: unknown };