# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for VibeGlobally Virtual Assistance Services
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /content/{section}:
    get:
      operationId: getSiteContent
      tags: [content]
      summary: Get site content for a section
      parameters:
        - in: path
          name: section
          required: true
          schema:
            type: string
            enum: [hero, services, results, tools, clients, team, values, contact, faq, privacy, terms, sampleCalls, facebookReview, logo, smtp]
      responses:
        "200":
          description: Site content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteContentResponse"
        "400":
          description: Unknown section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSiteContent
      tags: [content]
      summary: Update site content for a section (admin only)
      parameters:
        - in: path
          name: section
          required: true
          schema:
            type: string
            enum: [hero, services, results, tools, clients, team, values, contact, faq, privacy, terms, sampleCalls, facebookReview, logo]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Updated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteContentResponse"
        "400":
          description: Unknown section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SiteContentResponse:
      type: object
      properties:
        section:
          type: string
        content:
          type: object
          additionalProperties: true
      required:
        - section
        - content